move container job handler to /services/job_handlers
