+++ api/src/services/job_handlers/omnia_classic_azure_container_instances/__init__.py
@@ -23,7 +23,7 @@ AccessToken = namedtuple("AccessToken", ["token", "expires_on"])
 logging.getLogger("azure.mgmt").setLevel(logging.WARNING)
 logging.getLogger("azure.identity").setLevel(logging.WARNING)
 
-_SUPPORTED_TYPE = "DMT-Internal/DMT/AzureContainerInstanceJobClassic"
+_SUPPORTED_TYPE = ("DMT-Internal/DMT/AzureContainerInstanceJobClassic", "DMT-Internal/DMT/AzureContainerInstanceJob")
 
 
 def inject_environment_variables(template: dict, variables: List[EnvironmentVariable]) -> dict:
